Fix a timing issue on the v1.1 in TIMER_DMA mode
When reading from the ZuluSCSI v1.1 in PhyMode=2 (TIMER_DMA), the
ZuluSCSI will hit a  "start_dataInTransfer() timeout waiting
for previous to finish" error. This will hang a LC III and seems
to cause a data error when reading via dd on a Linux machine.

It seems that if there are around 16 sectors in the prefetch buffer
the above error occurs. Halving the PREFETCH_BUFFER from 8k to 4k
seems to resolve this issue without a noticeable speed reduction.

It should be noted that in it's default configuration with a
GreenPAK laid down, this does not occur. Thus there maybe a better
way to resolve this issue than limiting the PREFETCH_BUFFER across
the board for the ZuluSCSI v1.1.