RP2MCU: Add platform option RP2MCU_USE_CPU_PARITY
This is in preparation for 16-bit bus support.
With 16-bit bus, the data words are too wide for parity table.
Instead CPU based parity generation is used.
Its performance is much better on RP2350 Cortex-M33 than RP2040 Cortex-M0.

This commit does not yet implement the hardware acceleration
support for this.