Send message reject when requesting a wide bus
This is an attempt to fix issue: https://github.com/ZuluSCSI/ZuluSCSI-firmware/issues/542

The SCSI 2 spec has a weird negative (below), but
if we negate the meaning for a device that can only do 8bit transfers it
is assumed that the message response should be a MESSAGE REJECT message.

SCSI 2 spec
```
In addition, an SCSI device may initiate a WDTR message exchange whenever it is appropriate to negotiate a new
transfer width agreement. SCSI devices that are capable of wide data transfers (greater than eight bits) shall not
respond to a WDTR message with a MESSAGE REJECT message.
````

This has been implemented along with SCSI INQUIRY CDB checking the LUN
from the command instead of the LUN of the device.