ZuluSCSI Blaster use new variant from arduino-pico
Created a ZuluSCSI Blaster variant to the Rabbit Hole Computing custom
spin of the Arduino-Pico library. The reason is that Blaster uses
an RP2350B with the ability to use an RM2 so the rpipico2w variant
is not a proper match.

Added the PICO_RP2350A=0 define for the ZuluSCSI Wide target which is
how the RP2350B is differentiated from the A version. The Blaster
defines this in the variant included headers so it does not need
the define.