Backup ROM image before a changing flash layout
This build will backup a ROM image and warn the user if the backed up
ROM image will no longer fit in flash after transitioning to a new flash
layout in the upcoming firmware update.

The backed up filename extension is `.rom_bkup` and can be renamed to
`.rom` to reprogram the flash in the upcoming firmware update with
the same scsi id, block size and total file size as the original
programmed flash rom.