Reset on eject button in initiator mode
Because initiator mode on wide boards do not recover well on
disconnecting the device mode wide board it makes it hard to
test boards. The wide boards have a built in eject button.
To make testing easier, pressing the eject button resets the initiator
board after parking it in loop while waiting for the reset timer
to avoid resetting in the middle of an SD Card write.