Soft reset of initiator when eject button pressed
Using a soft reset of the initiator mode instead of an MCU reset.
The change now reruns the initiator init sequence then pauses
execution when the eject button is pressed. The LED pulsates to signify
it is paused. Pressing the eject button again starts the initiator
sequence again. Also switched to a debouncer to detect the eject button
release.