Fix messaging around firmware files on SD Card
This properly report to the user if a firmware file is found on the SD
card. If a correct firmware file is found to be read-only it reports
that. If the ZuluSCSI is powered on without and SD card and then
an SD card is inserted with proper firmware, the log reports that
the board needs to be powered cycled to install the firmware.

The USB interface is restarted in the init cycle. This is because
during a firmware update, the USB interface in Windows reports
that the USB device failed. After the firmware update the serial console
cannot be connected to. Restarting the USB interface solves this issue.