Fix onboard eject button on SD card reinsert
For boards with builtin eject buttons that default to the first
removable device when no device is assigned an eject button in
`zuluscsi.ini`, the eject button auto assignment is lost after
ejecting the SD card and reinserting.

Reinitializing the eject button's auto assignment after the SD card has
been reinserted fixes the issue.

Also cleaned up the log messaging when an eject button is pressed.