LED should breath in initiator mode
To immediately tell the difference between initiator mode and
target mode, the activity LED will "breath". That is it will
slowly go from one low brightness to a lower brightness and then
back again when the LED is in the off state.

In target mode the LED off state is simply off.

When in initiator MSC mode the LED breathes at a slower rate.
On the wide board in initiator state and the eject button
is pressed for the pause mode, the LED breathes at a rapid pace,
until the initiator state is reset by pressing the eject button again.