Fix unaligned buffer access in 16bit parity calc
It was found when using a different sector sizes that the 16 bit parity
calculation was crashing. Adding the CFSR and BFAR registers
for the RP2350 to the crash log helped track the issue down.

The cause was an unaligned buffer access, switching from the assembly
command ldm to two ldrs allows for the src buffer to be unaligned in
the case of using a sector size that causes unaligned data.

Co-authored-by: Petteri Aimonen <jpa@git.mail.kapsi.fi>