Controller and display working on RP2040
To get the controller and display code working on the RP2040, the
relevant UI objects are placed in the upper memory of the SCSI data
buffer (scsiDev.data). The length usable to the SCSI interface is
decreased as a result.
This lowers the transfer speed by about 5MB/s, from 15MB/s to 10MB/s.

For RP2350 based boards that use the controller code, an extra
buffer of reserved memory is used that acts just like the scsiDev.data
buffer so the SCSI buffer can remain its maximum size. This can be set
manually in a define as `ZULUSCSI_RESERVED_SRAM_LEN` but has a default
value if the controller code is active otherwise it is 0.

All the UI code has been placed in its own library
`lib/ZuluSCSI_UI_RP2MCU`. This helps with excluding the .text section
from the bootloader and placing it in Flash for RP2040 or SRAM
in the RP2350.

Also fixed some typos in the RP2350 linker script so now more
code should be run out of SRAM.