Make audio on RP2040 work again
Audio was initially disabled on the RP2040 target, now it uses the
same method for the controller board to assign memory from the
scsiDev.data buffer to the audio buffers. Because RP2040 audio and
the controller board share the same pin, only one can be used at a time.
If neither are used, the RP2040 can use the full size of the
scsiDev.data buffer for maximum throughput over the SCSI bus.

Currently the initial scsi map for the controller board is broken.
If the SD card is reinserted the map works properly.