Experimental code for dynamic buffer sizing
This contains `src/ZuluSCSI_buffer_control.cpp` and
`src/ZuluSCSI_buffer_control.h` which control the size of scsiDev.data
buffer usable to the SCSI interface. It can shrink the usable buffer
to allow other object and buffers to fit in the upper memory.

This will break certain things mentioned here:
https://github.com/ZuluSCSI/ZuluSCSI-firmware/pull/657#issuecomment-3269377617

But is saved in this branch in case something like this is needed later.