Wide support & memory alignment to buffer control
Add support for the lower 0-7 SCSI ID to use the Controller Interface
over I2C.

Various structs and objects need to be aligned to the nearest 32-bit
boundary. This addresses that issue and if the memory allocations
goes over the space in the buffer an assert is issue allowing
the zuluerr.txt file to point to exactly where the issue occurred.