Manage memory better with RP2040 & RP2350 boards
Summary:
 - Separate shard buffer used in the PR2040
 - ZuluSCSI RP2040 must be carefully tested with controller code for
 SRAM creep which may cause the shared buffer to fail at runtime
 - RP2350 controller code SRAM allocation is assigned at compile time
   avoiding the issue above with the RP2040
 - ZuluSCSI Pico W DaynaPORT build's buffers have all been reset to
 their default size due to the extra SRAM saved by an earlier commit

With the recent freeing up of SRAM on the RP2040 scsiDev.data no longer
needs to be shared.

Because of this a shared buffer that was used for the RP2350 boards
for loading the control board code has been utilized by the RP2040.
The reason for keeping shared memory is because the RP2040 S/PDIF audio
and controller board support are mutually exclusive and take up 10s of
kilobytes of SRAM, the down side is the size of the shared memory is
fixed at compile time so if the controller board code starts creep on
SRAM use the board will fail during runtime. The zuluerr.txt file
will contain a message on how many bytes over the shared buffer the
control code takes up.

To avoid this with the RP2350 boards, the shared buffer is turned off
and the controller board class objects are allocated at compile time.
This lets the controller code grow and if it is too large the compile
will fail.

ZuluSCSI DaynaPORT code no longer needs to adjust the buffers down,
because of the freed up SRAM via the linker script changes. The
SCSI/SD card interface should now run at full speed.