Fix splash screen and rotary response
The non-blocking splash screen was freezing when no SD card was inserted
and there is no ROM drive. It has been fixed by adding a controlLoop
into the do while statement that waits for the SD card to be inserted on
init.

The second issue was that during a removed SD card the timing
sensitivity of the rotary dial was being trampled on by the SD card
sense timeout. Because nothing is really happing during this time
running controlLoop for 500ms before attempting to retry the SD card
again allows the rotary to response to human input. This was added to
both the init wait loop when no SD card is detected at boot and in
the main loop after an SD card removal has been detected.

Also added the requirement that an SD card has to be present before
entering into MSC init and looping. This was generally avoided
because it is a zuluscsi.ini setting, but since rebooting into MSC
mode can happen while a card is not present it becomes an issue.

Finally some functions have been renamed to ones that make more sense.
 - splashScreenInit -> splashScreenPoll
 - deferredScreenLoad -> deferredChangeScreen