Additional functionality to USB serial menu
First off there is an actual menu printed to serial screen
describing the options. It activates when ever a key is pressed
that isn't one of the options.

Common options implemented in target mode and in MSC target mode:
    'R' - reboot device normally
    'M' - reboot with SD card as USB drive
    'I' - reboot with images presented as USB drives
    'D' - toggle debug, currently (on/off)
the "toggle debug" option states the actual state of logging debug info

When in MSC mode the following option is included
    'X' - exit from mass storage mode, eject USB drive(s)

Pressing 'y' after any of the options executes said option.

The current text should be proofed and the code could possibly
be folded into one menu to reduce redundancy.