Fix Blaster audio on reclock
ZuluSCSI Blaster audio is setup before reclocking because it is on by
default. Added the ability to reclock Blaster's I2S audio PIO
statemachine after the system clock has been reclocked. Which fixes the
issue.

Using different speed grades for the RP2040 now that audio is
compiled in and the default is 203.2 MHz. Speed grade A is now
251.2 MHz, B is 250, MHz and C is the original clock rate of 135.42 MHz
that S/PDIF was originally designed around.

Also changed the speed grade value of the Pico/Pico W. It is now
the same as the RP2040 but instead of C being 135.42 it is the
133 MHz.

Changed the ZuluSCSI RP2040 platformio.ini model define from
ZULUSCSI_V2_0 to ZULUSCSI_RP2040 to be more clear.