Fix issue with Blaster initiator mode being broken
This is a fix for issue: https://github.com/ZuluSCSI/ZuluSCSI-firmware/issues/688

The I2S audio state machine reclocking after and SD card is inserted was
causing issues with the initiator mode. As audio isn't required
for initiator mode, this fix disables the audio system when initiator
mode is enabled.