Get kiosk mode COW working on the v1.2
Decreasing the log buffer by half and gives enough memory for copy
on writes to work with the V1.2 which uses the V1.1+ firmware.

Also added a serial status logging for createImageFile().
It will display data written, data remaining, and the speed of the
writes and update every second on the same line on the serial console.
It does not write to the log buffer. There is a small though put penalty
for the status updates, but if the serial interface is not connected it
doesn't suffer from the slow down.