Fix eject buttons on RP2040 I2C lines
With the introduction of the control board and the merging of the
RP2040 Audio build into the base RP2040 build, eject buttons on the I2C
lines got lost.

This initiates GPIOs as inputs with pull up for buttons on the I2C lines
if EnableCDAudio is disabled (default state) and no control board is
found.