Fixing eject wrapping back to the first image
This is concerning issue: https://github.com/ZuluSCSI/ZuluSCSI-firmware/issues/705
In some cases of ejecting multi bin/cue directories with an eject button
, the images would not wrap back to the first image and would get stuck
indefinitely on a single image.

The cause was due to comparing a filename in the bin/cue directory
with the next possible disc name instead of the directory name of
the bin/cue disc.