Use splash screen and display create file progress
This returns the splash screen when no controller board is found but
a display exist. It also displays the version number on the error
screen. The error screen has the SD icon and it will indicate if
the SD card is present or not.

Because of the code path, the screen saver will trigger too.

Enabled all the progress bar displays too. Like the CREATE file,
ROM image copying, and kiosk modes. It returns to the error screen
after the progress bar finishes.