Put UI guards in place when OLED is not present (#712)
The guards against running UI elements when no control board (rotary
and buttons) were removed. This was done so if only the OLED was present
without the control board the UI elements like Create Image, and
Kiosk mode would be displayed.

Without any guards when neither the control board nor the OLED was
present the ZuluSCSI would crash when executing Create Image and other
unguarded procedures.

This fix added back the guards but checks for the OLED presence instead
of the control board.