Ignore ini system SCSI strings for network device (#714)
This is a requested solution for a discussion here
https://github.com/ZuluSCSI/ZuluSCSI-firmware/discussions/516

If vendor, product id, and revision SCSI string are set in
`zuluscsi.ini` under [SCSI] for system wide overrides, ignore those
values for a network device.

This is to avoid overwriting DaynaPORT SCSI strings. They can still
be overwritten in the device specific header like [SCSI4] in
`zuluscsi.ini`.