Small changes to SD card insertion and bus reset (#717)
Added more to closing of files and resetting state to
scsiDiskCloseSDCardImages. This removes the log message
"Already found an image directory, skipping..." when an SD card is
inserted after the first time.

Moved the keepCurrentImageOnBusReset check out of
cdromReinsertFirstImage and to where the bus reset code occurs. This
should free up cdromReinsertFirstImage to be used somewhere else if
necessary.